move state functions to getFirestoreData file
